+++ src/class/Config/Settings.php
@@ -324,6 +324,8 @@ class Settings extends Settings_Helper {
 			$debug_path = ABSPATH . WP_DEBUG_LOG;
 		}
 
+		$lang = str_replace( '-', '_', get_bloginfo( 'language' ) );
+
 		$imdb_admin_options = [
 			//--------------------------------------------------=[ Basic ]=--
 			'imdbplugindirectory_partial' => '/wp-content/plugins/lumiere-movies/',
@@ -343,7 +345,7 @@ class Settings extends Settings_Helper {
 			'imdbintotheposttheme'        => 'grey',
 			'imdblinkingkill'             => '0',
 			'imdbautopostwidget'          => '0',
-			'imdblanguage'                => get_bloginfo( 'language' ),
+			'imdblanguage'                => $lang,
 			'imdbdebug'                   => '0',                        /* Debug */
 			'imdbdebuglog'                => '0',                        /* Log debug */
 			/** @phpstan-ignore nullCoalesce.variable (PHPStan can't understand that WP_DEBUG_LOG is a const that can be string and bool) */